move remaining period dtype tests
